// angular-src/src/app/model/arts.ts
import { Time } from '@angular/common'

export interface ART{
    
        id: number
        artDescription: string
        artistId: number
        artistName: string
        AddedDate: Date

        artType: string
        thumbnail: string
        cover: string
        arts: string
        
        
}
export interface ARTMEIDA 
    {
        id: number
        artid: number
        src: string
         
    }